Migrate progress tracker functionality to mixin. Add file descriptor safe executor factory function.
